:alembic: Tidied up the color routing experiments
